+++ lib/arachni/browser_cluster/worker.rb
@@ -102,7 +102,7 @@ class Worker < Arachni::Browser
                 end
             end
         rescue TimeoutError => e
-            print_debug "Job timed-out after #{@job_timeout} seconds: #{@job}"
+            print_bad "Job timed-out after #{@job_timeout} seconds: #{@job}"
 
             # Could have left us with a broken browser.
             browser_respawn